% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromodal.R
\name{micromodal}
\alias{micromodal}
\title{Create a modal dialog}
\usage{
micromodal(id, title = NULL, content = NULL, footer = NULL)
}
\arguments{
\item{id}{A unique id for the modal}

\item{title}{The modal's title}

\item{content}{The modal's content}

\item{footer}{The modal's footer}
}
\value{
\code{\link[shiny:reexports]{shiny::tagList()}}
}
\description{
Create a modal dialog
}
