% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_micromodal.R
\name{use_micromodal}
\alias{use_micromodal}
\title{Use micromodal}
\usage{
use_micromodal()
}
\value{
\code{\link[htmltools:tagList]{htmltools::tagList()}}
}
\description{
Call this function once in your app's UI.
}
\details{
This function adds the dependencies needed for the modals.

See \code{\link[=micromodal]{micromodal()}} for a complete example.
}
\examples{
if (interactive()) {
  library(shiny)
  library(micromodal)

  ui <- fluidPage(
    use_micromodal(),
    # the rest of your UI code
  )
}
}
